chore: style 분리
