+++ src/hooks/useSlideObserver.ts
